Finalize delete domain functionality and refactor code.
